Type cast in stra.c compare

+++ stra.c
@@ -62,7 +62,7 @@ int Str_compare(const char s1[], const char s2[]) {
         i++;
     }
     
-    return s1[i] - s2[i];
+    return (int) s1[i] - s2[i];
 }
 
 /* A pointer to the first occurrence in the string pointed to by s1 of
